[memory64] Avoid using i64/BigInt for syscall return values. NFC (#16583)
The return values of syscalls are always errno value so there is no
need to using long/i64 for them.   This simplifies the generated JS
code a little and reduces the amount if i64 usage and the JS/wasm
boundary.

+++ system/lib/pthread/pthread_self_stub.c
@@ -15,7 +15,7 @@ uintptr_t __get_tp(void) {
 }
 
 // In case the stub syscall is not linked it
-static long dummy_getpid() {
+static int dummy_getpid() {
   return 42;
 }
 weak_alias(dummy_getpid, __syscall_getpid);
